refactor: move server-internal types from loon-types to loon-server
loon-types was exposing WAL, checkpoint, and control-object envelope
types to all downstream crates including loon-client, which has no
business accessing server internals. This tightens the API boundary:

- Move WAL types + codecs to loon-server::core::wal_types
- Move checkpoint types + codecs to loon-server::core::checkpoint_types
- Move control envelope types to loon-server::core::control_types
- Split object_store_keys: client keys stay in loon-types, server keys
  move to loon-server (eliminating a full copy-paste duplicate)
- Gate loon-server::core behind cfg(feature = "testkit") so only
  loon-testkit can reach internal modules
- Remove ciborium/zstd deps from loon-types (now only in loon-server)
- Replace glob re-exports with explicit exports in loon-types
