+++ crates/loon-server/Cargo.toml
@@ -5,9 +5,14 @@ edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
 
+[features]
+default = []
+testkit = []
+
 [dependencies]
 anyhow.workspace = true
 aws-sdk-s3.workspace = true
+ciborium.workspace = true
 http.workspace = true
 loon-types = { path = "../loon-types" }
 serde.workspace = true
@@ -15,6 +20,7 @@ serde_json.workspace = true
 sha2.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
+zstd.workspace = true
 
 [dev-dependencies]
 loon-testkit = { path = "../loon-testkit" }
