+++ crates/loon-testkit/tests/integration/checkpoint_builder_differential.rs
@@ -16,13 +16,13 @@ use loon_testkit::model::{
 };
 use loon_testkit::render::render_trace;
 use loon_testkit::scenario::Scenario;
-use loon_types::{
+use loon_server::core::checkpoint_types::{
     checkpoint_page_checksum_sha256, encode_checkpoint_manifest_json,
     encode_checkpoint_segment_envelope_zstd, CheckpointManifestEnvelope, CheckpointManifestPayload,
     CheckpointPage, CheckpointRow, CheckpointSegmentDescriptor, CheckpointSegmentEnvelope,
-    CheckpointSegmentPayload, CheckpointTableFamily, CheckpointTableManifest, FenceToken,
-    HeadState, InodeId, NamespaceId,
+    CheckpointSegmentPayload, CheckpointTableFamily, CheckpointTableManifest,
 };
+use loon_types::{FenceToken, HeadState, InodeId, NamespaceId};
 use serde::Deserialize;
 
 const TEST_WRITER_VERSION: &str = "loon-testkit-differential";
