+++ crates/loon-testkit/tests/integration/client_mutation_round_trip.rs
@@ -12,10 +12,8 @@ use loon_server::objectstore::fs::LocalFsStore;
 use loon_server::objectstore::keys::{namespace_head, namespace_lease};
 use loon_server::objectstore::ObjectStore;
 use loon_testkit::scenario::Scenario;
-use loon_types::{
-    ClientMutationResponse, ControlObjectKind, HeadState, HeadStateEnvelope, LeaseState,
-    LeaseStateEnvelope,
-};
+use loon_server::core::control_types::{ControlObjectKind, HeadStateEnvelope, LeaseStateEnvelope};
+use loon_types::{ClientMutationResponse, HeadState, LeaseState};
 use serde::Deserialize;
 use std::fs;
 use std::path::{Path, PathBuf};
