+++ crates/loon-types/Cargo.toml
@@ -6,9 +6,7 @@ license.workspace = true
 rust-version.workspace = true
 
 [dependencies]
-ciborium.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 sha2.workspace = true
 thiserror.workspace = true
-zstd.workspace = true
