+++ crates/loon-types/src/control.rs
@@ -1,17 +1,5 @@
-use crate::digest::sha256_hex;
 use crate::{ChangeSeq, FenceToken, InodeId, NamespaceId};
 use serde::{Deserialize, Serialize};
-
-pub const CONTROL_OBJECT_FORMAT_VERSION: u32 = 1;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum ControlObjectKind {
-    NamespaceHead,
-    NamespaceLease,
-    NamespaceProgress,
-    QueueShard,
-}
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct HeadState {
@@ -48,55 +36,4 @@ impl LeaseState {
     pub fn is_valid_at(&self, now_ms: u64) -> bool {
         self.lease_expires_at_ms > now_ms
     }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ProgressState {
-    pub namespace_id: NamespaceId,
-    pub work_class: String,
-    pub through_seq: ChangeSeq,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ControlObjectEnvelope<T> {
-    pub kind: ControlObjectKind,
-    pub format_version: u32,
-    pub writer_version: String,
-    pub payload_checksum_sha256: String,
-    pub state: T,
-}
-
-impl<T> ControlObjectEnvelope<T>
-where
-    T: Serialize,
-{
-    pub fn from_state(
-        kind: ControlObjectKind,
-        writer_version: impl Into<String>,
-        state: T,
-    ) -> Result<Self, serde_json::Error> {
-        Ok(Self {
-            kind,
-            format_version: CONTROL_OBJECT_FORMAT_VERSION,
-            writer_version: writer_version.into(),
-            payload_checksum_sha256: payload_checksum_sha256(&state)?,
-            state,
-        })
-    }
-
-    pub fn has_valid_payload_checksum(&self) -> Result<bool, serde_json::Error> {
-        Ok(self.payload_checksum_sha256 == payload_checksum_sha256(&self.state)?)
-    }
-}
-
-pub type HeadStateEnvelope = ControlObjectEnvelope<HeadState>;
-pub type LeaseStateEnvelope = ControlObjectEnvelope<LeaseState>;
-pub type ProgressStateEnvelope = ControlObjectEnvelope<ProgressState>;
-
-pub fn payload_checksum_sha256<T>(value: &T) -> Result<String, serde_json::Error>
-where
-    T: Serialize,
-{
-    let bytes = serde_json::to_vec(value)?;
-    Ok(sha256_hex(&bytes))
 }
