+++ crates/loon-types/src/lib.rs
@@ -7,7 +7,6 @@
 
 #![forbid(unsafe_code)]
 
-mod checkpoint;
 mod client;
 mod conflict;
 mod content;
@@ -17,18 +16,15 @@ mod ids;
 pub mod object_store;
 pub mod object_store_keys;
 pub mod server;
-mod wal;
 
-pub use checkpoint::*;
 pub use client::*;
 pub use conflict::*;
 pub use content::*;
-pub use control::*;
+pub use control::{HeadState, LeaseState};
 pub use digest::sha256_digest;
 pub use ids::*;
 pub use object_store::{ByteRange, ObjectMetadata, ObjectStore, ObjectStoreError, PutMode};
 pub use server::*;
-pub use wal::*;
 
 #[cfg(test)]
 mod tests;
